resolve some bug and add some missed features

+++ back-api/src/main/java/com/openclassrooms/mddapi/repositories/UserRepository.java
@@ -1,15 +1,20 @@
 package com.openclassrooms.mddapi.repositories;
 
 import com.openclassrooms.mddapi.models.UserEntity;
+import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+
 
     Optional<UserEntity> findByEmail(String email);
+
     Optional<UserEntity> findByUsername(String username);
 
 }
